+++ clients/frama-c-client/log-auditor/.gitignore
@@ -1,2 +1,1 @@
 LogAuditorIdentity.json
-log-files/*
