+++ clients/frama-c-client/log-creator/.gitignore
@@ -1,3 +1,2 @@
 LogCreatorIdentity.json
 Channel.json
-log-files/*
