+++ clients/frama-c-client/run-example/.gitignore
@@ -1,4 +1,3 @@
 LogAuditorIdentity.json
 LogCreatorIdentity.json
 Channel.json
-log-files/*
