+++ clients/frama-c-client/run-example/src/index.ts
@@ -4,7 +4,7 @@ import { createIdentity as createAuditorIdentity } from '../../log-auditor/src/c
 import { requestSubscription } from '../../log-auditor/src/request-subscription';
 import { authorizeSubscription } from '../../log-creator/src/authorize-subscription';
 import { writeStream } from '../../log-creator/src/write-stream-channel';
-import { getChannelData } from '../../log-auditor/src/get-channel-data';
+import { auditChannelData } from '../../log-auditor/src/audit-channel-data';
 import * as dotenv from 'dotenv';
 
 dotenv.config();
@@ -15,7 +15,7 @@ const run = async () => {
 	const subscriptionLink = (await requestSubscription(channelAddress)) as string;
 	await authorizeSubscription(channelAddress, subscriptionLink);
 	await writeStream(channelAddress);
-	await getChannelData(channelAddress);
+	await auditChannelData(channelAddress);
 };
 
 run();
